requirements.txt
Jinja2>=3.1,<4
Mastodon.py>=2.1,<3
python-dotenv>=1.0,<2
